WIP: continue refactoring - make tests run again
