+++ package.json
@@ -14,15 +14,15 @@
     "test:chunk:conductor": "bash ./scripts/run_test_chunk.sh 'not @Excluded and not @ExcludedOnMaster and not @OnlyRunsOnPNC and not @ExcludedOnConductor'",
     "test:chunk:nextUI": "NEXTUI=true bash ./scripts/run_test_chunk.sh 'not @Excluded and not @ExcludedOnMaster and not @OnlyRunsOnPNC and @NextUI'",
     "test:chunk": "bash ./scripts/run_test_chunk.sh 'not @Excluded and not @ExcludedOnMaster and not @OnlyRunsOnPNC' and @ExcludedOnNextUI",
-    "test:file": "./node_modules/.bin/cucumber-js --require steps --no-strict --exit --format @cucumber/pretty-formatter",
-    "test:must": "./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@Must' features",
-    "test:nextUI:file": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps --no-strict --exit --tags '@NextUI'",
-    "test:nextUI:must": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@Must and @NextUI' features",
-    "test:nextUI:preprod": "SKIP_PNC_VALIDATION=true NEXTUI=true ./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@PreProdTest and @NextUI and not @Excluded' features",
-    "test:nextUI": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@NextUI'",
-    "test:parallel": "curl -X DELETE http://localhost:3000/mocks && ./node_modules/.bin/cucumber-js --require steps --retry 0 --no-strict --exit --fail-fast --tags '@Parallel' --parallel $CONCURRENCY features",
-    "test:preprod": "SKIP_PNC_VALIDATION=true ./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@PreProdTest and not @Excluded' features",
-    "test": "./node_modules/.bin/cucumber-js --require steps --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags 'not @Excluded and not @ExcludedOnMaster and not @OnlyRunsOnPNC' features",
+    "test:file": "./node_modules/.bin/cucumber-js --require steps/index.js --no-strict --exit --format @cucumber/pretty-formatter",
+    "test:must": "./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@Must' features",
+    "test:nextUI:file": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps/index.js --no-strict --exit --tags '@NextUI'",
+    "test:nextUI:must": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@Must and @NextUI' features",
+    "test:nextUI:preprod": "SKIP_PNC_VALIDATION=true NEXTUI=true ./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@PreProdTest and @NextUI and not @Excluded' features",
+    "test:nextUI": "NEXTUI=true ./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@NextUI'",
+    "test:parallel": "curl -X DELETE http://localhost:3000/mocks && ./node_modules/.bin/cucumber-js --require steps/index.js --retry 0 --no-strict --exit --fail-fast --tags '@Parallel' --parallel $CONCURRENCY features",
+    "test:preprod": "SKIP_PNC_VALIDATION=true ./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags '@PreProdTest and not @Excluded' features",
+    "test": "./node_modules/.bin/cucumber-js --require steps/index.js --retry 5 --no-strict --exit --format @cucumber/pretty-formatter --fail-fast --tags 'not @Excluded and not @ExcludedOnMaster and not @OnlyRunsOnPNC' features",
     "update-deps": "ncu -u -x @cucumber/cucumber -x eslint && npm install",
     "prepare": "husky"
   },
