+++ steps/index.js
@@ -1,0 +1,8 @@
+require("./hooks");
+
+const { NEXTUI } = process.env;
+if (NEXTUI === "true") {
+  require("./next-ui");
+} else {
+  require("./legacy-ui");
+}
