--- steps/steps.js
@@ -1,6 +1,0 @@
-const { NEXTUI } = process.env;
-if (NEXTUI === "true") {
-  require("./steps.next-ui");
-} else {
-  require("./steps.legacy-ui");
-}
