+++ utils/auth.js
@@ -1,10 +1,10 @@
 const { expect } = require("expect");
 const jwt = require("jsonwebtoken");
 const { authType, timeout } = require("./config");
-const { login, authenticateUrl } = require("./urls.legacy-ui");
+const { login, authenticateUrl } = require("./urls");
 const dummyUsers = require("./dummyUserData");
 
-const nextui = () => process.env.NEXTUI === "true" || false;
+const nextui = process.env.NEXTUI === "true";
 
 const tokenSecret = () => process.env.TOKEN_SECRET || "OliverTwist";
 
@@ -47,7 +47,7 @@ const logInNormallyAs = async function (world, name) {
   await world.browser.clickAndWait("button[type='submit']");
   await page.waitForSelector('xpath/.//*[contains(text(), "Welcome ")]');
 
-  if (nextui()) {
+  if (nextui) {
     const [button] = await page.$$("xpath/.//a[contains(., 'Access New Bichard')]");
     if (button) {
       await Promise.all([button.click(), page.waitForNavigation()]);
@@ -93,7 +93,7 @@ const logInAs = async function (username) {
     await logInNormallyAs(this, username);
   }
 
-  const match = nextui() ? new RegExp(username, "i") : new RegExp(`You are logged in as: ${username}`, "i");
+  const match = nextui ? new RegExp(username, "i") : new RegExp(`You are logged in as: ${username}`, "i");
   expect(await this.browser.pageText()).toMatch(match);
 };
