--- utils/urls.legacy-ui.js
@@ -1,15 +1,0 @@
-const { config } = require("./config");
-
-const url = (path) => `${config.baseUrl}${path}`;
-
-const caseListPage = () => url("/bichard-ui/InitialRefreshList");
-const authenticateUrl = (token) => url(`/bichard-ui/Authenticate?token=${token}`);
-const logout = () => url("/bichard-ui/bichard-lo");
-const login = () => url("/users/login");
-
-module.exports = {
-  login,
-  caseListPage,
-  authenticateUrl,
-  logout
-};
